requirements.txt
pandas==2.2.3
zemberek_python==0.2.3
